RESTRUCTURED: project structure
