+++ PerseusMeteo/BusinessAPIs/OpenWeather/OpenWeatherModels.swift
