+++ PerseusMeteo/Configuration/Localization/ru.lproj/Localizable.strings
@@ -285,7 +285,7 @@
 
 "Label: Copyright Notice" = "Охраняется авторским правом © 7531 - 7534 Михаил A. Жигулин из Новосибирска.";
 
-"Label: Copyright Details" = "Летоисчесление ведется от Сотворения Мира в Звездном Храме в соответствии со славянским календарем в котором первый месяц сентябрь. Первое сентября начало 7534 года.";
+"Label: Copyright Details" = "Летоисчесление ведется от Сотворения Мира в Звездном Храме в соответствии со славянским календарем в котором первый месяц сентябрь. Первое сентября 2025 года от Р.Х. начало 7534 года.";
 
 "Label: Credits" = "Работы выполнены";
