+++ PerseusMeteo/Configuration/Translations/Translation_ru.plist
@@ -397,7 +397,7 @@
 	<key>Label: Copyright Notice</key>
 	<string>Охраняется авторским правом © 7531 - 7534 Михаил A. Жигулин из Новосибирска.</string>
 	<key>Label: Copyright Details</key>
-	<string>Летоисчесление ведется от Сотворения Мира в Звездном Храме в соответствии со славянским календарем в котором первый месяц сентябрь. Первое сентября начало 7534 года.</string>
+	<string>Летоисчесление ведется от Сотворения Мира в Звездном Храме в соответствии со славянским календарем в котором первый месяц сентябрь. Первое сентября 2025 года от Р.Х. начало 7534 года.</string>
 	<key>Label: Credits</key>
 	<string>Работы выполнены</string>
 	<key>Label: Balancing and Control</key>
