+++ PerseusMeteo/UIBusinessComponents/LocationView/SuggestionsViewItem.swift
