+++ PerseusMeteo/UIBusinessComponents/MeteoInformation/MeteoGroupData.swift
